profiling: Move CallStackAnalysis to the profiling plugin
CallStackAnalysis and its data provider, along with the view have been
moved to the profiling plugins.

Move the classes involving symbols and symbol providers to the profiling
plugins.

The Call Stack view is renamed to Flame Chart.

Those classes in tmf.core and tmf.ui have been deprecated

Change the swtbot tests that were using the CallStackView to
TimeGraphViewStub.

Add the profiling plugins to the feature.

Change-Id: I4b27edf2a63ee12d391c7895a655568b2c90d1f9
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/122818
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>