swtbot: Fix ResourcesViewTest.testFilter()
The filtered tree initialization triggers a delayed refresh job that can
interfere with the tree selection later in the test.

Start the test by typing in the filter text to trigger new refresh jobs
and wait for them to be complete. This will ensure no other refresh job
executes in the middle of the test.

Change-Id: I9b4d347a0723a0020bf3ffb2c8deb21b922ce8e9
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/122821
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>