profiling: Add the interfaces from incubator to profiling
The CallStackAnalysis implements both IFlameChartProvider and
ICallGraphProvider.

The segment store now comes with IFlameChartProvider instead of
callgraph, as callgraph can be used for sampling also, without segments.
The callgraph analysis is thus updated so that it is a sub-analysis of
CallStack instead of an analysis in itself.

The views are updated accordingly

Change-Id: I6eb12c68246608dc54825491711c4352116c7405
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/122974
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>