swtbot: Fix TimeGraphViewTest#testPalettes
Setting the palette in the presentation provider triggers a refresh that
queues a task to the UI thread and returns. There is a race condition in
the test, where the image was grabbed before the refresh actually
completed and the old colors were still present. It now uses a condition
to wait for the right colors to appear.

Change-Id: I4c7c023cf25fa0ead75f6a1302a8fb6a202321da
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/123007
Tested-by: CI Bot
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>