tmf: Bug 535034: AbstractSelectTreeViewer checked elements lost
Don't save the view context in updateTree(), instead save it in the
CheckStateListener or the filter text's ModifyListener.

Clear the filter text when the view context doesn't have one.

Clear the checked elements on traceOpened, traceSelected and reset. They
may be restored later.

Clear the filter text on reset.

Restore the pattern filter on traceOpened and traceSelected, to avoid
collapsing elements that do not match the old pattern filter, but
without changing the filter text which would erase the saved filter in
the view context, due to its ModifyListener.

Don't log a message when the trace does not have a data provider, it
might be an unsupported trace for the viewer that is opened.

In TriStateFilteredCheckboxTree, expand an element before propagating
the expanded state to its children in checkSubTree(), because if the
element had never been expanded, testFindItem won't find the element and
the state will not be set in the children. Then restore the element's
expanded state.

Change-Id: I4624f873f593adfa71a99392cf309a140da1b35e
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/123220
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/123939