criticalPath: Add a qualifier to edges
The state of the edges is an enum and cannot be changed by
implementation to add some data. A new string qualifier is added to the
edge, so implementers, for example, virtual machines, can add some
information to the link without changing its semantic.

That qualifier is displayed by the critical path view in the state.

Change-Id: I9bd4d7ac75934ff0e02a14d2e963be37d9a36a1b
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/123288
Tested-by: CI Bot
Reviewed-by: Hani Nemati <hani.nemati@gmail.com>
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>