tmf: Fix ProjectModelOutputTest
The method ProjectModelTestData.delayUntilTraceOpened() was called with
the TmfViewsElement and immediately returned without waiting.

The trace was left open after the tests. The @After method would then
delete the project, and the TmfProjectRegistry would later detect this
and close the corresponding traces in a deferred job, which could impact
subsequent tests if they had opened traces with the same resource.

The analysis's view output is checked in a wait condition instead of an
assert following a fixed delay.

Change-Id: I3e603acdc2f47300e4d682f563b25db1049d248b
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/123371
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>