tmf: Draw time graph event labels
Method ITimeEvent.getLabel() is added with default returning null. The
NamedTimeEvent class implements the method.

If the time event has a label, the time graph will now draw the label on
top of the state using a distinct color.

The drawing of label in overridden postDrawEvent() implementations is
removed when the time graph code can handle it instead.

Change-Id: If4c2b376016268ae665531f35b77faf871598c2b
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/125057
Tested-by: CI Bot
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Tested-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>