tmf.core: handle event request for aggregate traces
Previously the event processing was a one level operation. The event
could only be processed by the direct children of an experiment trace.
If the experiment is composed of a single trace and an experiment, the
single trace would be able to process the event but the event will not
be passed to the underlying traces of the experiment child. This patch
solves this issue so that aggregate traces could handle their
(grand)children trace's events.

Also add a state system event requested with a number of requested event

Change-Id: I6a5099a552932838c713b1c5505f82874b065cd2
Signed-off-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/125932
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>