xml: Add a class for state system path
Many components need a list of state attributes for a path in the state
system. Make one class handle it all.

This class also handles the #currentScenario special attribute that is
not really an attribute, but rather a way to specify the base quark to
use for the rest of the attributes.

Change-Id: I599587a065c92683dcfaa6c52188505d0b72b099
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/126131
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Jean-Christian Kouame <jean-christian.kouame@ericsson.com>