Bug 539092: tmf: Fix opening of experiment from views under experiments
commit 935c0a introduced this bug that opened the trace instead of the
experiment. This was caused because the delegate was not a
TmfAggregateAnalysisElement anymore as it was before. This patch
creates a new set of TmfAnalysisOutputElement with the aggregate
element as parent when the children of the analysis element are
requested.

Change-Id: I02398977d8e1db04490f1003243e477aa0721a36
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/129388
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot