lttng: bug 539636 Fix critical path with threaded IRQs
Distributed critical path requires machines to be synchronized with
network packets. Packets from ethernet network are received in softirq
contexts, but packets form wifi interfaces are received in threaded
IRQs, which means they are handled by an actual thread, that was waken
up outside the context of the hardware IRQ that caused it.

Additional events to kernel traces were required to detect the thread
wakeup, but also to associate a wakeup event with the packet that caused
it.

2 pairs of events are added:

* do_IRQ_entry/return: to wrap the full context of the IRQ, including
the hardware IRQ handler (irq_handler_entry/exit), the threaded IRQ
wakeup or the softirq that follows

* network_reception_entry/return: to wrap the context of packet
reception, so that the wakeup can be directly associated with the packet
that caused it and not be just another wakeup due to some userspace
operation that by chance happens right after a packet reception.

With those, the wakeup of the IRQ is linked to the hardware interrupt
that caused it, and the wakeup by packet reception can be automatically
matched by the origin of the packet, if available.

2 test traces are provided for unit test purposes. See the difference
with or without the new events.

To obtain a trace with lttng for this, in addition to the normal
critical path events, the following function in the kernel need to be
traced:

lttng enable-event -k --function do_IRQ do_IRQ
lttng enable-event -k --function netif_receive_skb_internal
                 netif_receive_skb_internal

Proper tracepoints should be added to the kernel first and to lttng
later in a near future.

Change-Id: Ie2f11fc7a9c225b58dab1944ff0d5d66f178523b
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/130503
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>