tmf: Make trace type selection use hint if valid
When selecting the trace type, if the provided trace type hint is valid,
return it immediately.

This prevents trace type selection from failing to find any valid trace
type if the trace type hint was overridden by a reduced candidate (that
could be a sub-class of the trace type provided by another plug-in) and
that reduced candidate had lowest confidence.

This also means that the trace type hint now has precedence over
candidate reduction. If the trace type hint is a generic trace type that
is valid, it will be used even if a more specific trace type is valid.

Change-Id: I2114eb4d918111c696d2004dc53eae1a2b87043a
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/130510
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>