ust: Fix symbol resolution when file not present
With LTTng 2.8+, the address is absolute and is usually resolved using
the lttng_ust_statedump:* events to get the binary files directly in the
file system. If the file is not present and we just want to add the
binary for GNU nm to work, symbols cannot be resolved.

This fixes this by having the UstDebugInfoSymbolProvider keep a list of
the other symbol providers for this trace and if the symbol cannot be
resolved, it will ask the other providers using the relative address
instead of the absolute one.

Change-Id: I3b2f7a4a4a3fa48e192a290b9600784595ee7125
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/130825
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>