tmf: Catch OperationCanceledException in TmfAbstractAnalysisModule
When an analysis is canceled (e.g. the trace was closed before the
analysis completed), an OperationCanceledException can be thrown, for
example by a call to SubMonitor.split() in the analysis code. This was
logged as an error and a failure dialog was shown to the user.

The exception is now caught silently and the normal actions are taken
for a canceled analysis.

Change-Id: I80a82afcf755805baf1fe99bdfe793ee2191d860
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/131414
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot