tmf: Fix TraceImportOperationTest
It appears that the destination folder can be null. It is suspected that
the resource change event that creates it is not always immediately
triggered or handled by the call to IFolder.create(). Force the refresh
of the Traces folder to ensure the destination folder is created, and
assert it is not null before continuing with the test.

Change-Id: I659340cb25914dee166451f1750cc249aee4d266
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/131689
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>