tmf: replace the synchronized "Stack" with "Deque"
Early classes of the Java API, such as Vector, Hashtable and
StringBuffer, were synchronized to make them thread-safe. Unfortunately,
synchronization has a big negative impact on performance, even when
using these collections from a single thread.

It is better to use their new unsynchronized replacements:

*    Deque instead of Stack

From JDK doc:

Deques can also be used as LIFO (Last-In-First-Out) stacks. This
interface should be used in preference to the legacy Stack class. When a
deque is used as a stack, elements are pushed and popped from the
beginning of the deque. Stack methods are precisely equivalent to Deque
methods as indicated in the table below:

Stack Method 	Equivalent Deque Method
push(e) 	addFirst(e)
pop() 	removeFirst()
peek() 	peekFirst()

Change-Id: Ic52308c9f337d95cf0c4070bd9881dc2cae4ef11
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/131776
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>