Support of dark theme in Trace Compass
In many UI code, the foreground and background is set manually.
Moreover, custom drawings are done that cannot be styled using a css
style, for example titles in SwtChart. To be able to customize the
colors for dark theme, the css file provided define color settings,
that are used at different locations. For time graphs, XY charts and
histogram the TimeGraphColorScheme is used to do the mapping.

The dark theme is not perfect right now, however it's better than
without it.

The following things can be improved in later patches:
- TmfEventsTable search highlighting
- TimeChart view foreground colors of event ticks
- State colors in Time Graph views based on theme
- Cursors in views (e.g. time graph view)
- FiltersView background

Change-Id: I0dbbd087f4821a7499dfb63f21a20d50ad77081e
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/132006
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>