tmf: Fix NullPointerException in TmfAggregateAnalysisElement
The default ITmfProjectModelElement.dispose() implementation gets the
children of the element and disposes all of them.

In TmfAggregateAnalysisElement the child analysis output elements are
copies of the real elements associated with one of the experiment
traces. They are created in the call to getChildren(). These copies do
not hold any resources so do not really need to be disposed. The real
analysis output elements will be disposed from their trace element.

When the TmfAggregateAnalysisElement is disposed, it therefore creates
new copies of the analysis output elements just before disposing them.
If the project is being deleted, the analysis output element constructor
can throw a NullPointerException because the element resource location
no longer exists.

Since there is no need to dispose anything, override dispose() in
TmfAggregateAnalysisElement to avoid unnecessarily creating elements.

Change-Id: Ie32ae02439c2239756d3aff000368fb4cf0207b4
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/132367
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>