swtbot: Fix XYDataProviderBaseTest.verifySeriesStyle()
In the UI thread, TmfCommonXAxisChartViewer.createSWTSeriesFromModel()
first creates the series, then sets the series style.

The tests wait for the series to be created (in the test thread) then
immediately verifies the series style (also in the test thread). At that
instant the series style may not have been set yet in the UI thread.

The verification method now does a syncExec() on the UI thread to ensure
that the runnable that created the new series has finished executing.

Change-Id: I768e293e970e1a7bc56fcda9fb03aad396d43ab6
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/133148
Tested-by: CI Bot
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>