swtbot: Fix TestRefreshTextTrace
Being the active editor does not guarantee that it is the active part
(it could be the active view).

Calling SWTBotEditor.setFocus() on the editor does not guarantee that it
becomes the active part. SWTBotEditor.show() should be called instead.

Otherwise the active part could be the Project Explorer view and then
the File > Refresh menu invokes the wrong action.

Change-Id: I51ca74f1b095877197e0e56a254ce8c9900a3d2b
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/134216
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>