ss: Add a 2d query starting from the end
Current 2d query parses state system nodes, one full level at a time
(breadth-first), which means intervals will be queried in no
particular general order, spanning the full range for each level.
That's good if intervals are processed independently, but not always the
case

This changes the default behavior to depth-first node iteration, so that
intervals with earlier start times will be processed first. This should
not change current behavior by much.

Sometimes, interval ranges may be important when processing. This patch
adds the possibility to specify a start time of the query range superior
to the end time, in which case, the nodes will be parsed depth-first
from the last one (so, typically, intervals ending last will be parsed
first).

In both cases, there is no sorting, but a general trend in interval
orders.

Change-Id: Iff376d301532f3c03663ee72c3e13f0f4067b88d
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/134665
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot