tmf: Fix time graph state color and height factor default handling
A state item in a time graph view's legend can have different default
color and/or height factor from the application in some cases. The
current code incorrectly detects this change of default value as if it
were a preference set by the user, and this prevents the default value
from the application to be used.

Do not set application default values as default in the preference
store. Instead, use a blank or invalid value to indicate that no user
preference is set.

Enable or disable the Reset button depending on whether a user
preference has been set or not.

Use setToDefault() to delete a user preference from the store when the
Reset button is pressed.

Add StateItem.getStateHeightFactor() helper method.

Change-Id: Ie1abea0139d39a66b997ccef18ac4d7189d1fe3a
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/136165
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>