dataprovider: Fix DataProviderParameterUtils to accept list of numbers
In some case the parameters can contain a list both Integer and Long.
Now when we transform the list to Long list we support any numbers that
has a longValue. Example, when deserialize Jackson will take a json
number and will create the "smallest" number (Integer or Long) and can
create a List with both inside.

Change-Id: Ifc8d4ea0e83a8541f38d74b830d98cacbe09adc9
Signed-off-by: Simon Delisle <simon.delisle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/137019
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>