lttng: Bug 544022: Remove Virtual Machine analysis and view
It was deprecated in Trace Compass 3.2 and moved to the Trace Compass
Incubator project, where it is being actively developed. The
functionality in Trace Compass has thus been marked as deprecated,
inviting the user to get the feature from the Incubator. The Incubator's
update site can be found
http://download.eclipse.org/tracecompass.incubator/master/repository/.

Deprecation patch:
https://git.eclipse.org/r/#/c/106386/

Change-Id: Id90ab8ba19d2096bf9eb343d3b6bdc0b55f9050d
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/138773
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>