ss: Bug 546343: IllegalStateException in ThreadedHistoryBackend
Allow intervals with start time == Long.MIN_VALUE and attribute != -1.
This can happen if the trace is empty, and should not be considered as a
malformed "poison pill".

When a trace has no events, use Long.MIN_VALUE as the state system
history end time, to match the start time.

Change-Id: Ic6f8cb885c46c723e7871df16c7532bc10a0523c
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/140451
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot