model: Move IFilterableDataModel in IElementResolver
Deprecates IFilterableDataModel and move its content to
IElementResolver.

These 2 classes serve similar purposes: allow various elements to
provide data on which to search and filter. One is a multimap to allow
an element to have multiple values for a single key, while the other did
not allow that. Only multimap remains now. The map method is deprecated.

The filters have been updated to supported multimap as inputs.

Change-Id: I2e258adc1df22a3f4274907630e7788abc995635
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/140707
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: CI Bot