tmf.ui: Fixes to base data provider view
It should not be assumed that the entries without parent represent the
trace and have no value. It is just how it has been implemented so far.
The TraceEntry class represents the root entries and keep the origin data
provider for all its children. It is for the models to tell if they have
row models or not.

Also, events without a value may still have a label. In that case, they
should be created as NamedTimeEvent instead of NullTimeEvents.

Change-Id: I4d8ee846b9eee59d2ea7f36bd93084f988f22021
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/141465
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>