timing: Filter segment on the content's field as well
The Content aspect will resolve the full content as a string. Also add
each of the content's field as separate field/value for filter input to
allow filtering on those.

This moves the aspect resolving method as a static method of
ISegmentStoreProvider.

It also uses the new Multimap predicate for the segment table

Change-Id: I40c72384bbc9e6edacb788bff73bbff82f883121
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/141593
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>