ctf.core: Add additional search for CPU aspect.
CPU aspects may be in fields if they aren't defined in the streams.
This is a fall-back. Performance impact should be low if the trace
has stream based CPU info like LTTng, and others should have more
information.

This changes the behaviour, now the CPU will be blank rather than
display -1 for undefined.

Change-Id: Ibdda099c5f93e62498d2c99554d3393a6ae61329
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/141800
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: CI Bot