tmf: Make time scale format use Time Format preference by default
By default, AbstractTimeGraphView, TmfXYChartViewer, Histogram and
TimeChartView now set the time scale format according to the value of
the Time Format 'Date and Time format' preference.

If the preference is set to 'TTT (secs in epoch)', then
TimeFormat.RELATIVE is used. In all other cases, TimeFormat.CALENDAR is
used, where the displayed format is determined by the current window
range, not by the specific 'Date and Time format' preference.

This default behavior can be changed for a specific subclass
implementation of AbstractTimeGraphView or TmfXYChartViewer by calling
setTimeFormat(TimeFormat). Calling the method with null restores the
default behavior.

AbstractTimeGraphView subclass implementations that used to set the time
format directly on the TimeGraphViewer should now call the method on the
view instead.

Change-Id: I27210e07e57936b513308e94300f6420d4f511c8
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/142691
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>