tmf: Add ToolTipString to tool tip handler
The category, name and value of tool tip items can now optionally be
instances of the new class ToolTipString.

The existing method to add plain text name and value is undeprecated,
and a new method is added to add plain text category, name and value.

Static methods in ToolTipString allow creation of a tool tip string from
a plain text string, an HTML string, or a timestamp, which creates a
time hyperlink.

When converting between plain text and HTML string, the HTML entities
are escaped/unescaped.

Change-Id: I85838ec47e8c259f5e6ca1f47a0e7f031ec572fe
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/142765
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>