tmf: Improve tooltip layout and preferred size calculations
Make each tooltip content calculate once and return its preferred size
at creation.

Only BrowserContent needs to include the vertical scroll bar width in
the preferred size. Its ScrolledComposite does not need V_SCROLL style,
since it is handled by the Browser widget.

DefaultContent should calculate the preferred size using the inner
content Composite, because the ScrolledComposite preferred size includes
its scroll bars' sizes, which are not visible if the tool tip is at the
correct preferred size.

The BrowserContent body margin can be configured, it is reduced to 3px.

The BrowserContent content has a 1px margin, it should be included in
the preferred size calculations.

The DefaultContent does not need to override its parent layout, which is
a FillLayout. The GridLayout added a default marginWidth/Height of 5.

Change-Id: I6547ea914cd8afba75793ad8f0ac43ced1375d7f
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/143892
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>