swtbot: Fix ConditionHelpers
Rename some static methods to start with a lower-case letter.

Fix viewIsClosed() to check that the view is closed, instead of checking
that the view is not active.

Rename viewIsOpened() to viewIsActive(), which is what the condition
checks. It is not possible to get a SWTBotView for a view that is not
open.

Add failure message for viewIsClosed() and viewIsActive().

Fix Javadoc @return tag for methods that return a condition.

Change-Id: I02103c4ede05a92ae30e43f54f13928e3120af6d
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/144397
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>