tmf: Validate element under trace folder in TmfOpenTraceHelper
When opening an experiment trace, if the underlying trace has been
removed from the file system in the background, the call to
TmfTraceElement.getElementUnderTraceFolder() will return itself, e.g.
the trace under the experiment element.

In that case the OpenProjectElementJob can later fail with a strange
error message (e.g. 'Path must be a valid directory' for a CTF trace),
because the experiment trace's resource (a dummy empty file) is being
used.

The code now checks for this condition early and fails with the message
'Trace <element path> not found'.

Change-Id: I17b3b075355c5f972a5e5b126c92d485fc11f630
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/144582
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>