releng: Activate build for 4.12 target by default
This patch also stores the tracing.product file that is needed to
build for older platforms (e.g. 4.11 or older) as well. Since there is
no easy way to copy this file from maven before the maven reactor
dependencies are parsed, the copy operation has to be done manually
before building with maven (e.g. mvn clean install). Using Jenkins
this can be done automatically when building for older platforms.

Here are the steps to build for platforms (e.g. 4.11) from the
repository root directory:
  cd rcp/org.eclipse.tracecompass.rcp.product
  cp legacy/tracing.product .
  cd ../../
  mvn clean install -Dtarget-platform=tracecompass-e4.11

Here are the steps to build for the current or later platforms
(>= 4.12):
  mvn clean install

Change-Id: I7c0514d2bf5198654d7c01e7cca978b88ee179f9
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/145110
Tested-by: CI Bot
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>