common: Improve formatted string parsing and tests
Format#parseObject(String, ParsePosition) should update the parse
position up to the last character that was parsed in the string. The
various common formatter tests use the base test class to test different
test cases with complete/incomplete strings.

Also the DecimalUnitFormatTest now also uses the FormatTestBase class as a
base, which allows to remove the DecimalUnitFormatErrorTest class.

Change-Id: Ic95d8515e25ad1955d0da8412870fbe85a23343b
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/145554
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>