swtbot: Fix ProjectExplorerTracesFolderTest.getTraceProperty()
The Properties view contains a PageBook. When the workbench selection
changes, a new page can be shown and the previous page is hidden.

When trying to get the Resource properties from the tree found in the
Properties view (when the selection comes from a trace element in the
Project Explorer view), we have to make sure to get the right tree. If
we get a reference to the tree before the selection has changed, we
could use a reference to the previous tree that gets hidden (for
example, the tree showing trace event properties when the selection
comes from the event table). This tree will never contain the Resource
properties, even if the Properties view shows them (from a different
tree in a different PageBook page).

Wait until the view contains a tree that contains the Resource
properties before getting a reference to the tree and continuing to get
the property value from this tree.

Change-Id: I9bb6ee88841d88f5954a136b0adbfa5402cd2f13
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/146024
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>