releng: Fix sonar build introduced due to SonarQube version updates
- Update sonar-maven-plugin version to 3.6.0.1398
- Update jacoco-maven-plugin version to 0.8.4
- Ignore ITmfEventField.java for sonar due to checkstyle bug
Checkstyle 4.19 fails parsing method signatures with null annotations
and ellipsis, for example:
ITmfEventField getField(String @NonNull ... path);
- SonarQube deprecated 'sonar.jacoco.reportPath' in favor to
'sonar.coverage.jacoco.xmlReportPaths' (see [1])
This lead to that the code coverage was not populated properly to
SonarQube because the xml path was used by SonarQube and not the
jacoco.exec file. Now, it's necessary to add the option jacoco:report
and the location of the jacoco.dataFile to the maven command line:
mvn Djacoco.dataFile=../../target/jacoco.exec jacoco:report sonar:sonar
With this and the changes in the pom.xml file, an xml report is created
for each test plug-in using the jacoco.exec file, that is appended by
each test plug-in. See [2] for email thread on the SonarQube community
mailing list. Note that JaCoCo Plugin 1.0.2 or later need to be
installed on SonarQube.

[1] https://jira.sonarsource.com/projects/MMF/issues/MMF-1651?filter=allissues
[2] https://community.sonarsource.com/t/jacoco-coverage-switch-from-deprecated-binary-to-xml-format-in-a-tycho-build-shows-0/10917

Change-Id: If1f8761363427ebdd7f15e4a2b136ee2bd329fee
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/145010
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>