filters: Do not re-generate filter expressions at each test
For large traces, a lot of time was spent on the 'prepareValue' method
because for each test, all the FilterExpressions are re-generated
because they needed to be inserted in a queue and they were consumed.
Now they are generated once in an immutable list and there is no queue
consumption when testing values.

[Fix] Performance improvement on filter expressions

Change-Id: I5c4f57e341933ed601ab25b3b152a2cbdcbb231e
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/146818
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>