timing: Filter segments table in a job
Filtering segments can be a time consuming task. It was often done in the
UI thread, in the context of a signal manager. When filtering is
requested, it should be done in a job and the view updated when it is
done, to not block the UI thread or signal manager.

Change-Id: Id51aab9d13b22b4355535f00432b80fb14b0daf2
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/146819
Tested-by: CI Bot
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>