tmf: Bug 550454: Fix invalid memory allocation in base data provider
If a trace in an experiment doesn't have events, then the zoomEntries()
will be called with start -INFINITE and experiment end time. This will
lead to a state system query from -INFINITE to end time which can bust
the memory or cause illegal allocations depending on the experiment
start time.

[Fixed] Bug 550454: OOME when adding an empty trace to an experiment

Change-Id: Ib1e5fa3e21c3ad4f49065019cc88ee839c0fe834
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/147856
Tested-by: CI Bot
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>