tmf.core: Introduce Callsite analysis
Paves the way to have code lookup in all time graphs when available.
To have the trace work on this, it needs to have events with callsites,
a cpu aspect and a trace with a host id.

Note: The default implementation only takes into account the top-most
callsite, as it is the one that is "executing" at that moment.

[Added] Callsite analysis

Change-Id: Ic5fe50dd9f17b54c7b4bf8dfb29635b684a7ae53
Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/148606
Tested-by: CI Bot