tmf: Improve TimeGraphLegend layout
Use a single column of equally layed-out groups. Make the link group the
last of these groups.

Initialize each group to use 1 column of state items. If the dialog
shell is taller than the available area, initialize each group to use 2
columns of state items instead.

When the dialog is resized, compute the highest number of columns (up to
a maximum of 8) of state items to use that fits in the new dialog width.

Remove an extra unnecessary composite.

Adjust the margins of the group layouts.

[Changed] Improved time graph legend layout with dynamic columns

Change-Id: I1aec64b3053b163b3d9e565bed089d00695d8600
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/148840
Tested-by: CI Bot
Reviewed-by: Simon Delisle <simon.delisle@ericsson.com>
Tested-by: Simon Delisle <simon.delisle@ericsson.com>